+++ CHANGELOG.md
@@ -15,10 +15,11 @@
 * Right click option for recipe and package treeview item to copy its path
 * [#13](https://github.com/afri-bit/vsconan/issues/13) Managing editable packages
   * List editable packages in the treeview
-  * Remove editable package via GUI
+  * Remove editable package via Treeview
   * Open editable package in VS Code
   * Open editable package in Explorer
   * Copy editable path to clipboard
+  * Remove editable package via command and quickpick (simple option)
 
 ### Changed
 * The configuration for extension is migrated to official VS Code `settings.json`. Custom global `config.json` under `~/.vsconan` is now **deprecated**.